\name{analyzePeaks-package}
\alias{analyzePeaks-package}
\alias{analyzePeaks}
\docType{package}
\title{
\packageTitle{analyzePeaks}
}
\description{
\packageDescription{analyzePeaks}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{analyzePeaks}
\packageIndices{analyzePeaks}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{analyzePeaks}

Maintainer: \packageMaintainer{analyzePeaks}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
