% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeaks.R
\name{getPeaks}
\alias{getPeaks}
\title{Extract meaningfull peaks from spectra}
\usage{
getPeaks(spectra, model, calibMethod = "TIC", noiseMethod = "SuperSmoother",
  tolerance = 2e-04, minFrequency = 0.25, SNR = 2, halfWindowSize = 5)
}
\arguments{
\item{spectra}{list of \code{MALDIquant} \code{\link{AbstractMassObject}})}

\item{model}{\code{\link{glmnet}} model to deal with}
}
\value{
matrix of peaks intensities suitable for \code{\link{predict}}
}
\description{
Extract meaningfull peaks from spectra
}
