% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeaks.R
\name{getPrediction}
\alias{getPrediction}
\title{Calculate Predictions}
\usage{
getPrediction(peaks, model)
}
\arguments{
\item{peaks}{matrix from \code{\link{getPeaks}}}

\item{model}{model \code{\link{glmnet}} model to deal with}
}
\value{
list with values: class, probability of class I, probability of class II
}
\description{
Calculate Predictions
}
