% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeaks.R
\name{getReference}
\alias{getReference}
\title{Prepare reference MassPeaks for warping of the input}
\usage{
getReference(model)
}
\arguments{
\item{model}{\code{\link{glmnet}} model to deal with}
}
\value{
\code{\link{MassPeaks}} object
}
\description{
Prepare reference MassPeaks for warping of the input
}
